bring in ariadne and try to get it connected to mysql for preliminary serving data, having problems with docker build
