+++ .gitignore
@@ -1,1 +1,2 @@
 .env
+/__pycache__
