+++ auth_helper.py
@@ -1,12 +1,11 @@
-import json
-from six.moves.urllib.request import urlopen
+from flask import Flask, request, _request_ctx_stack
 from functools import wraps
-from flask import Flask, request, jsonify, _request_ctx_stack
-from flask_cors import cross_origin
-from jose import jwt
-from flask_mysqldb import MySQL
+from six.moves.urllib.request import urlopen
 from dotenv import load_dotenv
 import os
+import json
+from jose import jwt
+
 
 load_dotenv()
 
@@ -14,27 +13,11 @@ AUTH0_DOMAIN = os.getenv('AUTH0_DOMAIN')
 API_AUDIENCE = os.getenv('AUTH0_AUDIENCE')
 ALGORITHMS = ["RS256"]
 
-APP = Flask(__name__)
-
-APP.config['MYSQL_HOST'] = os.getenv('MYSQL_HOST')
-APP.config['MYSQL_USER'] = os.getenv('MYSQL_USER')
-APP.config['MYSQL_PASSWORD'] = os.getenv('MYSQL_PASSWORD')
-APP.config['MYSQL_DB'] = os.getenv('MYSQL_DB')
-APP.config['MYSQL_PORT'] = os.getenv('MYSQL_PORT')
-
-mysql = MySQL(APP)
-
 # Error handler
 class AuthError(Exception):
     def __init__(self, error, status_code):
         self.error = error
         self.status_code = status_code
-
-@APP.errorhandler(AuthError)
-def handle_auth_error(ex):
-    response = jsonify(ex.error)
-    response.status_code = ex.status_code
-    return response
 
 def get_token_auth_header():
     """Obtains the Access Token from the Authorization Header
@@ -111,35 +94,3 @@ def requires_auth(f):
         raise AuthError({"code": "invalid_header",
                         "description": "Unable to find appropriate key"}, 401)
     return decorated
-
-def query(sql,options):
-    cur = mysql.connection.cursor()
-    cur.execute(sql,(options))
-    mysql.connection.commit()
-    data = jsonify(cur.fetchall())
-    cur.close()
-    return data
-
-
-@APP.route("/api/somequery/")
-def getcamps():
-    sql = "Select * from camps"
-    return query(sql,())
-
-@APP.route("/")
-def HELLO():
-    return "This is a landing page"
-# This doesn't need authentication
-@APP.route("/api/public")
-@cross_origin(origin = "localhost",headers=["Content-Type", "Authorization"])
-def public():
-    response = "Hello from a public endpoint! You don't need to be authenticated to see this."
-    return jsonify(message=response)
-
-# This needs authentication
-@APP.route("/api/private")
-@cross_origin(origin="localhost", headers=["Content-Type", "Authorization"])
-@requires_auth
-def private():
-    response = "Hello from a private endpoint! You need to be authenticated to see this."
-    return jsonify(message=response)
