+++ db_helper.py
@@ -1,0 +1,7 @@
+def query(mysql, sql,options):
+    cur = mysql.connection.cursor()
+    cur.execute(sql,(options))
+    mysql.connection.commit()
+    data = cur.fetchall()
+    cur.close()
+    return data
