+++ type_defs.py
@@ -1,0 +1,7 @@
+from ariadne import gql
+
+type_defs = gql("""
+    type Query {
+        hello: String!
+    }
+""")
